requirements.txt
flask
flask-cors
ultralytics
opencv-python
PyYAML 